+++ kernel/pseudo_devices.hpp
@@ -6,12 +6,12 @@ public:
     inline ssize_t read(char* buf, size_t at_most_bytes) override
     {
         bzero(buf, at_most_bytes);
-        return static_cast<ssize_t>(at_most_bytes);    // validated by FileDescription
+        return static_cast<ssize_t>(at_most_bytes);  // validated by FileDescription
 
     }
     inline ssize_t write(char*, size_t bytes) override
     {
-        return bytes;
+        return static_cast<ssize_t>(bytes);  // validated by FileDescription
     }
 };
